Update Homepage with recent blog post
